+++ apps/backend/src/app/api/v1/contact-channels/crud.tsx
@@ -129,11 +129,28 @@ export const contactChannelsCrudHandlers = createLazyProxy(() => createCrudHandl
     }
 
     const updatedContactChannel = await prismaClient.$transaction(async (tx) => {
-      await ensureContactChannelExists(tx, {
+      const existingContactChannel = await ensureContactChannelExists(tx, {
         projectId: auth.project.id,
         userId: params.user_id,
         contactChannelId: params.contact_channel_id || throwErr("Missing contact channel id"),
       });
+
+      // if usedForAuth is set to true, make sure no other account uses this channel for auth
+      if (data.used_for_auth) {
+        const existingWithSameChannel = await tx.contactChannel.findUnique({
+          where: {
+            projectId_type_value_usedForAuth: {
+              projectId: auth.project.id,
+              type: data.type !== undefined ? crudContactChannelTypeToPrisma(data.type) : existingContactChannel.type,
+              value: data.value !== undefined ? data.value : existingContactChannel.value,
+              usedForAuth: 'TRUE',
+            },
+          },
+        });
+        if (existingWithSameChannel && existingWithSameChannel.id !== existingContactChannel.id) {
+          throw new KnownErrors.ContactChannelAlreadyUsedForAuthBySomeoneElse(data.type ?? prismaContactChannelTypeToCrud(existingContactChannel.type));
+        }
+      }
 
       if (data.is_primary) {
         // mark all other channels as not primary
@@ -215,3 +232,12 @@ export const contactChannelsCrudHandlers = createLazyProxy(() => createCrudHandl
     };
   }
 }));
+
+
+function crudContactChannelTypeToPrisma(type: "email") {
+  return typedToUppercase(type);
+}
+
+function prismaContactChannelTypeToCrud(type: "EMAIL") {
+  return typedToLowercase(type);
+}