+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack-backend
 
+## 2.6.28
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.28
+  - @stackframe/stack-emails@2.6.28
+
+## 2.6.27
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.27
+  - @stackframe/stack-emails@2.6.27
+
 ## 2.6.26
 
 ### Patch Changes