+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/e2e-tests
 
+## 2.6.28
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.28
+
+## 2.6.27
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.27
+
 ## 2.6.26
 
 ### Patch Changes