+++ examples/cjs-test/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/example-cjs-test
 
+## 2.6.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.28
+
+## 2.6.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.27
+
 ## 2.6.26
 
 ### Patch Changes