+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @stackframe/init-stack
 
+## 2.6.28
+
+## 2.6.27
+
 ## 2.6.26
 
 ### Patch Changes