+++ packages/stack-emails/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/stack-emails
 
+## 2.6.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.28
+  - @stackframe/stack-ui@2.6.28
+
+## 2.6.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.27
+  - @stackframe/stack-ui@2.6.27
+
 ## 2.6.26
 
 ### Patch Changes