+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @stackframe/stack-sc
 
+## 2.6.28
+
+## 2.6.27
+
 ## 2.6.26
 
 ### Patch Changes