+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/stack-shared
 
+## 2.6.28
+
+### Patch Changes
+
+- Bugfixes
+  - @stackframe/stack-sc@2.6.28
+
+## 2.6.27
+
+### Patch Changes
+
+- Bugfixes
+  - @stackframe/stack-sc@2.6.27
+
 ## 2.6.26
 
 ### Patch Changes