+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @stackframe/stack
 
+## 2.6.28
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.28
+  - @stackframe/stack-ui@2.6.28
+  - @stackframe/stack-sc@2.6.28
+
+## 2.6.27
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.27
+  - @stackframe/stack-ui@2.6.27
+  - @stackframe/stack-sc@2.6.27
+
 ## 2.6.26
 
 ### Patch Changes